chore(edgegw): Add schema/types (#404)
