+++ internal/provider/edgegw/app_port_profile_resource.go
@@ -1,4 +1,3 @@
-// Package network provides a Terraform resource.
 package edgegw
 
 import (
@@ -12,20 +11,10 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 
 	"github.com/hashicorp/terraform-plugin-framework/resource"
-	schemaR "github.com/hashicorp/terraform-plugin-framework/resource/schema"
-
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/setplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
-
-	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-
-	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
 
 	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/client"
 	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/provider/common/org"
@@ -53,21 +42,6 @@ func NewPortProfilesResource() resource.Resource {
 type portProfilesResource struct {
 	client *client.CloudAvenue
 	org    org.Org
-}
-
-type portProfilesResourceModel struct {
-	ID          types.String `tfsdk:"id"`
-	Name        types.String `tfsdk:"name"`
-	VDC         types.String `tfsdk:"vdc"`
-	Description types.String `tfsdk:"description"`
-	AppPorts    types.List   `tfsdk:"app_ports"`
-}
-
-type portProfilesResourceModelAppPorts []portProfilesResourceModelAppPort
-
-type portProfilesResourceModelAppPort struct {
-	Protocol types.String `tfsdk:"protocol"`
-	Ports    types.Set    `tfsdk:"ports"`
 }
 
 func (rm *portProfilesResourceModel) AppPortsFromPlan(ctx context.Context) (appPorts portProfilesResourceModelAppPorts, diags diag.Diagnostics) {
@@ -159,7 +133,7 @@ func (r *portProfilesResource) Metadata(_ context.Context, req resource.Metadata
 
 // Schema defines the schema for the resource.
 func (r *portProfilesResource) Schema(ctx context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
-	resp.Schema = r.superSchema(ctx).GetResource(ctx)
+	resp.Schema = portProfilesSchema(ctx).GetResource(ctx)
 }
 
 func (r *portProfilesResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
@@ -455,88 +429,4 @@ func (r *portProfilesResource) ImportState(ctx context.Context, req resource.Imp
 
 	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("vdc"), vdcID)...)
 	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("name"), idParts[1])...)
-}
-
-func (r *portProfilesResource) superSchema(_ context.Context) superschema.Schema {
-	return superschema.Schema{
-		Resource: superschema.SchemaDetails{
-			MarkdownDescription: "Provides a NSX-T App Port Profile resource",
-		},
-		Attributes: map[string]superschema.Attribute{
-			"id": superschema.StringAttribute{
-				Common: &schemaR.StringAttribute{
-					MarkdownDescription: "The ID of the VM.",
-					Computed:            true,
-				},
-				Resource: &schemaR.StringAttribute{
-					PlanModifiers: []planmodifier.String{
-						stringplanmodifier.UseStateForUnknown(),
-					},
-				},
-			},
-			"name": superschema.StringAttribute{
-				Common: &schemaR.StringAttribute{
-					MarkdownDescription: "Application Port Profile name.",
-				},
-				Resource: &schemaR.StringAttribute{
-					Required: true,
-				},
-			},
-			"description": superschema.StringAttribute{
-				Common: &schemaR.StringAttribute{
-					MarkdownDescription: "Application Port Profile description.",
-				},
-				Resource: &schemaR.StringAttribute{
-					Optional: true,
-					PlanModifiers: []planmodifier.String{
-						stringplanmodifier.UseStateForUnknown(),
-					},
-				},
-			},
-			"vdc": superschema.StringAttribute{
-				Common: &schemaR.StringAttribute{
-					MarkdownDescription: "ID of VDC or VDC Group",
-				},
-				Resource: &schemaR.StringAttribute{
-					Required: true,
-					PlanModifiers: []planmodifier.String{
-						stringplanmodifier.RequiresReplace(),
-						stringplanmodifier.UseStateForUnknown(),
-					},
-				},
-			},
-			"app_ports": superschema.ListNestedAttribute{
-				Common: &schemaR.ListNestedAttribute{
-					MarkdownDescription: "List of application ports.",
-					Required:            true,
-				},
-				Attributes: map[string]superschema.Attribute{
-					"ports": superschema.SetAttribute{
-						Common: &schemaR.SetAttribute{
-							MarkdownDescription: "Set of ports or ranges.",
-							Computed:            true,
-						},
-						Resource: &schemaR.SetAttribute{
-							Optional:    true,
-							ElementType: types.StringType,
-							PlanModifiers: []planmodifier.Set{
-								setplanmodifier.UseStateForUnknown(),
-							},
-						},
-					},
-					"protocol": superschema.StringAttribute{
-						Common: &schemaR.StringAttribute{
-							MarkdownDescription: "Protocol.",
-						},
-						Resource: &schemaR.StringAttribute{
-							Required: true,
-							Validators: []validator.String{
-								stringvalidator.OneOf("ICMPv4", "ICMPv6", "TCP", "UDP"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
 }
