+++ internal/provider/edgegw/app_port_profile_schema.go
@@ -1,0 +1,102 @@
+package edgegw
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+
+	schemaR "github.com/hashicorp/terraform-plugin-framework/resource/schema"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/setplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+
+	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
+)
+
+func portProfilesSchema(_ context.Context) superschema.Schema {
+	return superschema.Schema{
+		Resource: superschema.SchemaDetails{
+			MarkdownDescription: "Provides a NSX-T App Port Profile resource",
+		},
+		Attributes: map[string]superschema.Attribute{
+			"id": superschema.StringAttribute{
+				Common: &schemaR.StringAttribute{
+					MarkdownDescription: "The ID of the VM.",
+					Computed:            true,
+				},
+				Resource: &schemaR.StringAttribute{
+					PlanModifiers: []planmodifier.String{
+						stringplanmodifier.UseStateForUnknown(),
+					},
+				},
+			},
+			"name": superschema.StringAttribute{
+				Common: &schemaR.StringAttribute{
+					MarkdownDescription: "Application Port Profile name.",
+				},
+				Resource: &schemaR.StringAttribute{
+					Required: true,
+				},
+			},
+			"description": superschema.StringAttribute{
+				Common: &schemaR.StringAttribute{
+					MarkdownDescription: "Application Port Profile description.",
+				},
+				Resource: &schemaR.StringAttribute{
+					Optional: true,
+					PlanModifiers: []planmodifier.String{
+						stringplanmodifier.UseStateForUnknown(),
+					},
+				},
+			},
+			"vdc": superschema.StringAttribute{
+				Common: &schemaR.StringAttribute{
+					MarkdownDescription: "ID of VDC or VDC Group",
+				},
+				Resource: &schemaR.StringAttribute{
+					Required: true,
+					PlanModifiers: []planmodifier.String{
+						stringplanmodifier.RequiresReplace(),
+						stringplanmodifier.UseStateForUnknown(),
+					},
+				},
+			},
+			"app_ports": superschema.ListNestedAttribute{
+				Common: &schemaR.ListNestedAttribute{
+					MarkdownDescription: "List of application ports.",
+					Required:            true,
+				},
+				Attributes: map[string]superschema.Attribute{
+					"ports": superschema.SetAttribute{
+						Common: &schemaR.SetAttribute{
+							MarkdownDescription: "Set of ports or ranges.",
+							Computed:            true,
+						},
+						Resource: &schemaR.SetAttribute{
+							Optional:    true,
+							ElementType: types.StringType,
+							PlanModifiers: []planmodifier.Set{
+								setplanmodifier.UseStateForUnknown(),
+							},
+						},
+					},
+					"protocol": superschema.StringAttribute{
+						Common: &schemaR.StringAttribute{
+							MarkdownDescription: "Protocol.",
+						},
+						Resource: &schemaR.StringAttribute{
+							Required: true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("ICMPv4", "ICMPv6", "TCP", "UDP"),
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+}
